蓝盾流水线自动打准备发布灰度tag - gray/v1.8.35-rc-2022-01-17

+++ src/backend/ci/ext/tencent/stream/biz-stream-tencent/src/main/kotlin/com/tencent/devops/stream/trigger/parsers/PipelineDelete.kt
@@ -174,10 +174,17 @@ class PipelineDelete @Autowired constructor(
                 !streamPipelineBranchService.hasBranchExist(gitRequestEvent.gitProjectId, pipelineId)) {
                 logger.info("event: ${gitRequestEvent.id} delete file: $filePath with pipeline: $pipelineId ")
                 gitPipelineResourceDao.deleteByPipelineId(dslContext, pipelineId)
-                processClient.delete(
-                    gitRequestEvent.userId, gitProjectConf.projectCode!!, pipelineId,
-                    channelCode
+                val pipelineInfoResult = processClient.getPipelineInfo(
+                    projectId = gitProjectConf.projectCode!!,
+                    pipelineId = pipelineId,
+                    channelCode = channelCode
                 )
+                if (pipelineInfoResult.data != null) {
+                    processClient.delete(
+                        gitRequestEvent.userId, gitProjectConf.projectCode!!, pipelineId,
+                        channelCode
+                    )
+                }
                 // 删除相关的构建记录
                 gitCIEventService.deletePipelineBuildHistory(setOf(pipelineId))
             }
